pyproject.toml
[project]
name = "latent-assets"
version = "0.1.0"
description = "Images, embeddings and search."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "accelerate>=1.12.0",
    "einops>=0.8.2",
    "fastapi[standard]>=0.128.7",
    "jupyterlab>=4.1.6",
    "pillow",
    "python-multipart>=0.0.22",
    "sentence-transformers>=5.2.2",
    "timm>=1.0.24",
    "torch==2.9.0+cu130",
    "transformers<5",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu130"

